Implement undo/redo history branching
This enables multiple redo futures,
a feature available in editors such as vim and emacs.
Undo branches / undo tree allows you to undo some changes,
then make a new change, while keeping all changes available in the undo tree.
The undo tree can therefore prevent awkward scenarios such as:
 * Losing all redo futures by accidentally pressing a button during undo
  * Copying the file or save it with git when trying multiple
  alternatives

It's implemented by modifying the edit stack elements to keep track of
all possible futures, instead of throwing alternative futures away.
A new event has been added, notifying all movement in the tree.

The method moveTo is used to move to a certain node in the tree, by
providing its index (a unique and ordered number). This index is used
to find the common ancestor of the current node and the target node.
The undo/redo path through the common ancestor is the shortest number of
transformations needed to go from the current node to the target.

The history tree is visualized in a contrib plugin on two overlayed
canvases, to enable some future optimizations.
The node position calculation of the visualization is cached,
but several more calculations have further optimization potential.